requirements.txt
accelerate==1.7.0
av
bitsandbytes
datasets>=2.16.0
deepdiff
elasticsearch==8.0.0
fastapi
fsspec
huggingface_hub
ibm_cloud_sdk_core
json_repair
jsonref
langchain_core
langchain_ibm
langchain_ollama
librosa
loguru
matplotlib
modelscope
numpy
omegaconf
openai
optimum
packaging
pandas
pathos
peft
Pillow
pyarrow
pydantic
pytest
python-dotenv
PyYAML
regex
Requests
rouge
rouge_score
sacrebleu
safetensors
seaborn
sentence_transformers
setuptools
sqlglot
textstat
torch
tqdm
transformers
trl==0.19.1
typing_extensions
urllib3
vllm